requirements.txt
camelot-py==0.11.0
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
cryptography==41.0.1
et-xmlfile==1.1.0
ghostscript==0.7
numpy==1.24.3
opencv-python==4.7.0.72
openpyxl==3.1.2
packaging==23.1
pandas==2.0.2
pdfminer.six==20221105
pdftopng==0.2.3
Pillow==9.5.0
pycparser==2.21
pycryptodome==3.18.0
pypdf==3.11.0
pytesseract==0.3.10
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
tabulate==0.9.0
typing_extensions==4.6.3
tzdata==2023.3
